fallos/pyproject.toml
[project]
name = "fallos"
version = "0.1.0"
description = "Microservicio para inyección de fallos y ataques"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.26.0"
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0"
]